requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.1
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.13
langchain-text-splitters==0.2.2
langsmith==0.1.82
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.35.7
orjson==3.10.5
packaging==24.1
portalocker==2.10.0
protobuf==5.27.2
pydantic==2.8.0
pydantic_core==2.20.0
PyJWT==2.8.0
pymongo==4.8.0
pypdf==4.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
qdrant-client==1.9.2
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.4.2
tiktoken==0.7.0
tqdm==4.66.4
twilio==9.2.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
